% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{compvar}
\alias{compvar}
\title{Variance comparison between two independent samples}
\usage{
compvar(x, y, test = "dos-colas")
}
\arguments{
\item{x}{A vector or array for the first sample}

\item{y}{A vector or array for the second sample}

\item{test}{A character value containing one of the following option: "dos-colas", "inferior", "superior"}
}
\value{
Return a list containing the statistic's value and the P-value for the test
}
\description{
Compare de variance of two independent samples, depending of the alternative hypothesis
}
\examples{
x<-c(10.8,11.1,10.4,10.1,11.3)
y<-c(10.8,10.5,11,10.9,10.8,10.7,10.8)
compvar(x,y,test="superior")
}
\references{
Conover, W. J. (1999) "Practical Nonparametric Statistics", 3rd Edition, Jhon Wiley & Sons.Inc
}
