% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{monregnp}
\alias{monregnp}
\title{Monotone no parametric regresion}
\usage{
monregnp(data)
}
\arguments{
\item{data}{a vector, array or data frame}
}
\value{
Ni idea que retorna
}
\description{
description here!!!!
}
\references{
librito de no parametrica
}
