% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{cochranq.test}
\alias{cochranq.test}
\title{Cochranq Hyphotesys test for matrix}
\usage{
cochranq.test(mat)
}
\arguments{
\item{mat}{a matrix object}
}
\value{
the Cochranq hyphotesys test
}
\description{
description here!!!!
}
\references{
librito de no parametrica
}
