% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{CramerV}
\alias{CramerV}
\title{Cramer V}
\usage{
CramerV(x, y = NULL, digits = 4, ...)
}
\arguments{
\item{x}{first element}

\item{y}{second element}

\item{digits}{number of digits you want in the result}

\item{...}{Another things that idk}
}
\value{
a numeric value indicating the rao crammer v
}
\description{
description here!!!!
}
\references{
librito de no parametrica
}
